Correct the code

+++ Principes/Principes/Open_Closed/Correct/Rectangle.cs
@@ -1,11 +1,14 @@
 ﻿namespace Principes.Open_Closed.Correct
 {
+    /// <summary>
+    /// Derived class extend class 'Shape'
+    /// </summary>
     public class Rectangle : Shape
     {
         public int Height { get; set; }
         public int Width { get; set; }
 
-        public override int Area()
+        public override double Area()
         {
             return Height * Width;
         }
