+++ Principes/Principes/Open_Closed/Correct/Shape.cs
@@ -1,7 +1,10 @@
 ﻿namespace Principes.Open_Closed.Correct
 {
+    /// <summary>
+    /// Abstract class for extension
+    /// </summary>
     public abstract class Shape
     {
-        public abstract int Area();
+        public abstract double Area();
     }
 }
