+++ Principes/Principes/Open_Closed/Correct/Square.cs
@@ -1,10 +1,13 @@
 ﻿namespace Principes.Open_Closed.Correct
 {
+    /// <summary>
+    /// Derived class extend class 'Shape'
+    /// </summary>
     public class Square : Shape
     {
         public int Sides;
 
-        public override int Area()
+        public override double Area()
         {
             return Sides * Sides;
         }
